fix(Entity classes): Rename entity package to entities
